requirements.txt
aniso8601==9.0.1
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
db-dtypes==1.0.5
dnspython==2.3.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-api-core==2.11.0
google-auth==2.16.3
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.8.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mod-wsgi @ file:///D:/python/perch-mount-flask-app/mod_wsgi-4.9.2-cp39-cp39-win_amd64.whl
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
proto-plus==1.22.2
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.7.0
pymongo==4.3.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
